# Conflicts:
#	app/views/api/v1/communities/delete_tour_stop_v1.json.jbuilder

+++ app/models/latch.rb
@@ -27,6 +27,7 @@ class Latch < ApplicationRecord
                     notify_pusher({error: "Invalid File"})
                 end
             rescue => exception
+                puts "--------------------------- an error occured -----------------------------"
                 notify_pusher({error: "Error in parsing " + "the uploaded CSV" + " file."})
             end
         ensure
@@ -35,26 +36,26 @@ class Latch < ApplicationRecord
     end
 
     def save_lock_info(community, lock_data)
-        lock_name = lock_data[1]; lock_id = lock_data[2]; stop_data = lock_data[3];
-        data = parse_stop(stop_data)
-        lock = community.latch.latch_locks.find_by(lock_id: lock_id) rescue nil
+        lock_name = lock_data[1]; lock_id = lock_data[2]; stop_data = lock_data[3]; door_uuid = lock_data[4];
 
-        if lock_name.present? and lock_id.present? 
+        data = parse_stop(stop_data)
+        lock = community.latch.latch_locks.find_by(door_uuid: door_uuid) rescue nil
+
+        if lock_name.present? and lock_id.present? and door_uuid.present?
             if lock.nil? and data.nil?
-                LatchLock.create(lock_id: lock_id, lock_name: lock_name, latch_id: community.latch.id)
+                LatchLock.create(door_uuid: door_uuid, lock_name: lock_name, latch_id: community.latch.id, lock_id: lock_id)
             elsif lock.present? and data.nil?
                 lock.update_attributes(lock_id: lock_id, lock_name: lock_name, latch_id: community.latch.id)
             else  # (lock.nil? and data.present?) or (lock.present? and data.present?)
-                data.latch_locks.destroy_all
+                data.latch_locks.map{|l| l.update_attributes(stop_type: nil, stop_id: nil)}
                 lock.destroy if lock.present?
-                data.latch_locks.create(lock_id:lock_id, lock_name: lock_name, latch_id: community.latch.id)
+                data.latch_locks.create(door_uuid: door_uuid, lock_name: lock_name, latch_id: community.latch.id, lock_id:lock_id)
             end
         end
     end
 
 
     def parse_stop(stop_name)
-      
         data=nil
 
         if stop_name.include?('-')
@@ -63,14 +64,31 @@ class Latch < ApplicationRecord
             building_name = nil
             unit_name = stop_name
         end
-       
-        data = community.units.where('(marketing_name = ? or provider_unit_id = ?) and building = ?', unit_name, unit_name, building_name).first rescue nil
-        data = community.units.where('marketing_name = ? or provider_unit_id = ?', stop_name, stop_name).first rescue nil unless data.present?
-        
+
+        if building_name == nil
+            building_name_ = ''
+        else
+            building_name_ = building_name
+        end
+
+        data = community.units.where('(marketing_name = ? or provider_unit_id = ?) and (building = ? or building = ?)', unit_name, unit_name, building_name, building_name_).first rescue nil
+        data = community.units.where('(marketing_name = ? or provider_unit_id = ?) and (building = ? or building = ?)', stop_name, stop_name, nil, '').first rescue nil unless data.present?
+
         data = community.amenities.where(name: stop_name).first if data.nil?
         data = community.elevators.where(name: stop_name).first if data.nil?
         data = community.building_starting_point.where(name: stop_name).first if data.nil?
         data = community.tour.name == stop_name ? community.tour : nil if data.nil?
+
+        if data.nil? and stop_name.count('-') > 1
+            (stop_name.count('-')+1).times.each do |i|
+                if i >= 2
+                    building = (stop_name.split('-', i+1).first ((stop_name.split('-', i+1).length) -1)).join('-')
+                    unit_name = stop_name.split('-', i+1).last
+                    data = community.units.where('(marketing_name = ? or provider_unit_id = ?) and building = ?', unit_name, unit_name, building).first rescue nil
+                end
+                break if data.present?
+            end
+        end
 
         return data
     end
